pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61", "setuptools-scm"]

[project]
name = "sbatcher"
authors = [{name = "Yuji Kanagawa", email = "yuji.kanagawa@oist.jp"}]
description = "A simple templating tool for executing sbatch"
license = {text = "Apache2"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Typing :: Typed",
]
requires-python = ">= 3.7"
dependencies = [
    "click >= 8.0",
    "jinja2 >= 3.0",
    "pyserde[toml] >= 0.8",
    "typer >= 0.6",
]
dynamic = ["version"]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
homepage = "https://github.com/kngwyu/sbatcher"
bug-tracker = "https://github.com/kngwyu/sbatcher/issues"

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]

[tool.setuptools.dynamic]
version = {attr = "sbatcher.__version__"}

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.isort]
profile = "black"

[tool.pyright]
extraPaths = [
    ".nox/tests-3-10/lib/python3.10/site-packages/",
    "src"
]